[CHORE] Layout and pages refactored (#50)
* refactored layout code

* made pages more leaner
